tools: enable commenting out lines of image conversion configuration

+++ remove-converted-images.sh
@@ -10,17 +10,19 @@ read -n 1
 set -e
 
 for conversion in $(cat image-conversions.conf); do
-  scope=$(echo "$conversion" | cut -d: -f1 -)
-  original=$(echo "$conversion" | cut -d: -f2 -)
-  destination=$(echo "$conversion" | cut -d: -f3 -)
-  size=$(echo "$conversion" | cut -d: -f4 -)
-  echo "Cleaning up $scope::$original"
-  
-  for f in source/$scope/$original; do
-    filename=$(basename "$f")
-    filebase="${filename%.*}"
-    destfile="source/$scope/$destination$filebase.png"
-    echo "  Removing $destfile"
-    rm -f $destfile
-  done
+  if [ $(echo "$conversion" | cut -c1) != "#" ]; then
+    scope=$(echo "$conversion" | cut -d: -f1 -)
+    original=$(echo "$conversion" | cut -d: -f2 -)
+    destination=$(echo "$conversion" | cut -d: -f3 -)
+    size=$(echo "$conversion" | cut -d: -f4 -)
+    echo "Cleaning up $scope::$original"
+    
+    for f in source/$scope/$original; do
+        filename=$(basename "$f")
+        filebase="${filename%.*}"
+        destfile="source/$scope/$destination$filebase.png"
+        echo "  Removing $destfile"
+        rm -f $destfile
+    done
+  fi
 done
